+++ packages/giu-examples/pages/material.js
@@ -24,7 +24,6 @@ import {
   Hints,
   flexContainer,
   flexItem,
-  boxWithShadow,
   TextInput,
   ColorInput,
   COLORS,
@@ -240,7 +239,7 @@ const StyleUtilsExample = () => (
       <FlexSpacer />
       <span>Right</span>
     </div>
-    <div style={boxWithShadow({ padding: 3 })}>A box with a shadow</div>
+    <div className="giu-box-shadow" style={{ padding: 3 }}>A box with a shadow</div>
   </div>
 );